% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_join.R
\name{difference_join}
\alias{difference_join}
\title{Join two tables based on absolute distance between their columns}
\usage{
difference_join(x, y, by = NULL, max_dist = 1)
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{by}{Columns by which to join the two tables}

\item{max_dist}{Maximum distance to use for joining}
}
\description{
Join two tables based on absolute distance between their columns
}
\examples{

library(dplyr)

head(iris)
sepal_lengths <- data_frame(Sepal.Length = c(5, 6, 7), Type = 1:3)

iris \%>\%
  difference_join(sepal_lengths, max_dist = .5)

}
\seealso{
\code{\link{stringdist-methods}}
}

