% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_join.R
\name{fuzzy_join}
\alias{fuzzy_join}
\title{Join two tables based not on exact matches, but rather with a function
describing whether two vectors are matched or not}
\usage{
fuzzy_join(x, y, by = NULL, match_fun, ...)
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{by}{Columns of each to join}

\item{match_fun}{Vectorized function given two columns, returning
TRUE or FALSE as to whether they are a match}

\item{...}{Extra arguments passed to match_fun}
}
\description{
The \code{match_fun} argument is called once on the pair of all unique comparisons:
thus, it should be efficient and vectorized.
}

