% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_join.R
\name{regex_join}
\alias{regex_join}
\title{Join two tables based on a regular expression in one column
matching the other}
\usage{
regex_join(x, y, by = NULL)
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{by}{Columns by which to join the two tables}
}
\description{
Join a table with a string column by a regular expression column
in another table
}
\examples{

library(dplyr)
library(ggplot2)
data(diamonds)

diamonds <- tbl_df(diamonds)

d <- data_frame(regex_name = c("^Idea", "mium", "Good"),
                type = 1:3)

# When they are inner_joined, only Good<->Good matches
diamonds \%>\%
  inner_join(d, by = c(cut = "regex_name"))

# but we can regex match them
diamonds \%>\%
 regex_join(d, by = c(cut = "regex_name"))

}
\seealso{
\code{\link{str_detect}}
}

