% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdist_join.R
\name{stringdist_inner_join}
\alias{stringdist_inner_join}
\title{Join two tables based on fuzzy string matching of their columns}
\usage{
stringdist_inner_join(x, y, by = NULL, max_dist = 2, ignore_case = FALSE,
  ...)
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{by}{Columns by which to join the two tables}

\item{max_dist}{Maximum distance to use for joining}

\item{ignore_case}{Whether to be case insensitive (default yes)}

\item{...}{Arguments passed on to \code{\link{stringdist}},
most notably "method": see \code{stringdist-methods} in the
stringdist package.}
}
\description{
Join two tables based on fuzzy string matching of their columns.
This is useful, for example, in matching free-form inputs in
a survey or online form, where it can catch misspellings and
small personal changes.
}
\examples{

library(dplyr)
library(ggplot2)
data(diamonds)

d <- data_frame(approximate_name = c("Idea", "Premiums", "Premioom",
                                     "VeryGood", "VeryGood", "Faiir"),
                type = 1:6)

# no matches when they are inner-joined:
diamonds \%>\%
  inner_join(d, by = c(cut = "approximate_name"))

# but we can match when they're fuzzy joined
diamonds \%>\%
 stringdist_inner_join(d, by = c(cut = "approximate_name"))

}

